fatura güncellemedeki hata düzeltildi.

+++ Backend/src/main/java/com/softwareProject/banksApplication/dto/request/invoice/InvoiceUpdateRequest.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -15,6 +16,6 @@ public class InvoiceUpdateRequest {
     private Long id;
     private Long invoiceNo;
     private InvoiceInfo.InvoiceType invoiceType;
-    private int invoiceAmount;
+    private BigDecimal invoiceAmount;
     private boolean autobill;
 }
